third_party/blink/web_tests/platform/generic/svg/foreignObject/fO-percentage-height-style-expected.txt

Test use of percentages inside foreignObject

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS getComputedStyle(nestedSvg).height is '250px'
PASS getComputedStyle(nestedSvg).height is '125px'
PASS getComputedStyle(nestedSvg).height is '200px'
PASS successfullyParsed is true

TEST COMPLETE
